// src/frontend/libs/openapi/out/model/configOption.ts
/**
 * IBSYS RETAIL MANAGER API
 * Cette API propose plusieurs points de terminaison pour gérer les opérations de vente au détail, telles que la création et la mise à jour de produits, la gestion des stocks et le traitement des commandes.
 *
 * The version of the OpenAPI document: x.x.x
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ConfigOption { 
    /**
     * Énumération des options de configuration.
     */
    key?: ConfigOption.KeyEnum;
    value?: string;
}
export namespace ConfigOption {
    export type KeyEnum = 'LAST_INVOICE_NUMBER' | 'LAST_BL_NUMBER';
    export const KeyEnum = {
        InvoiceNumber: 'LAST_INVOICE_NUMBER' as KeyEnum,
        BlNumber: 'LAST_BL_NUMBER' as KeyEnum
    };
}